# Move log_normal to Aten(CPU)