@ezyang 2020-01-08
It looks like this change may have resulted in numeric differences. Internal test:

```
AssertionError: 
Arrays are not almost equal to 6 decimals

Mismatch: 100%
Max absolute difference: 1.8119812e-05
Max relative difference: 3.180625e-06
 x: array([5.696916], dtype=float32)
 y: array([5.696934], dtype=float32)
```

I'm rerunning to check if it's flakiness.

The test contents:

```
	
    @given(
        data=st.lists(st.floats(min_value=0.0, max_value=1.0), min_size=1, max_size=20),
        eps=st.floats(min_value=1e-6, max_value=1e-2),
    )
    @multi_env()
    def test_logit(self, data, eps):
        def compare_reference(input, eps, output):
            input = np.clip(input, a_min=eps, a_max=1.0 - eps)
            expect = np.log(input / (1.0 - input))

            np.testing.assert_array_almost_equal(
                expect, blob_to_numpy(output), decimal=6
            )

        input_arr = np.array(data).astype(np.float32)
        input = as_blob(input_arr)

        # Test with default eps
        output = Logit()(input)
        compare_reference(input_arr, 1e-6, output)

        # Test with customized eps
        output = Logit(eps=eps)(input)
        compare_reference(input_arr, eps, output)
```

and Logit:

```
class Logit: ...
    def forward_pytorch(self, input):
        eps = 1e-6 if self._eps is None else self._eps

        input_clipped = torch.clamp(input, min=eps, max=1.0 - eps)
        return torch.log(input_clipped / (1.0 - input_clipped))
```