@XiaobingSuper 2020-01-10
@ezyang , I can't reproduce the failed case, by the way, the failed case call torch.clamp and torch.log, which not call in this changed function.